examples/aiohttp/dogs/requirements.txt
--extra-index-url https://test.pypi.org/simple/
redis
tartiflette-aiohttp
